requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1

# Database and vector storage
supabase-py==2.0.0
pgvector==0.2.3
sqlalchemy==2.0.23

# AI and ML
openai==1.3.5
anthropic==0.5.0
langchain==0.0.335
sentence-transformers==2.2.2

# Cross-platform
streamlit==1.28.2
pywebview==4.4.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.0
flake8==6.1.0 